re-enabled D7AActP

+++ stack/modules/alp/alp_layer.h
@@ -128,7 +128,7 @@ void alp_layer_process_d7ap_result(uint8_t* alp_command, uint8_t alp_command_len
  * \param alp_command   The raw command
  * \param alp_command_length The length of the command
  */
-void alp_layer_process_d7aactp(d7ap_session_config_t* config, uint8_t* alp_command, uint8_t alp_command_length);
+void alp_layer_process_d7aactp(d7ap_session_config_t* config, uint8_t* alp_command, uint32_t alp_command_length);
 
 /*!
  * \brief Process the ALP command and output the result on the console.
